requirements.txt
python-dotenv
streamlit
pandas
duckdb
plotly
numpy
